+++ stringpool.hpp
@@ -10,21 +10,21 @@ public:
 
   /* Adds string `s` to the stringpool. If `force_append` is true, no attempts
    * are made to return an existing string in the string pool. */
-  size_t add(const char* s, bool force_append = false);
+  int add(const char* s, bool force_append = false);
 
   /* A pool is optimized when it it sorted by length (descending). */
   bool isOptimized() const;
 
   /* Return the number of NUL terminated strings */
-  size_t count() const;
+  int count() const;
 
   /* Returns the ID for string `s`.
    * If the string is empty or it could not be found, 0 is returned. */
-  size_t find(const char *s) const;
+  int find(const char *s) const;
 
-  inline const char*  get(size_t id)   const { return storage.c_str() + id; }
-  inline size_t       size()           const { return storage.size();       }
-  inline size_t       capacity()       const { return storage.capacity();   }
+  inline const char*  get(int id)      const { return storage.c_str() + id; }
+  inline int          size()           const { return storage.size();       }
+  inline int          capacity()       const { return storage.capacity();   }
   inline void         resize(size_t n)       { storage.resize(n);           }
   inline void         reserve(size_t n)      { storage.reserve(n);          }
   inline char*        data()                 { return const_cast<char*>(storage.data()); }
